imports + agon core integration + centrality heuristic

+++ agon/src/main/java/fr/univ/bordeaux/application/ai/heuristics/MixedHeuristic.java
@@ -1,9 +1,11 @@
 package fr.univ.bordeaux.application.ai.heuristics;
 
+import fr.univ.bordeaux.agonCore.bitboard.AgonBoard;
+
 public class MixedHeuristic implements Heuristics {
 
     @Override
-    public double evaluate(AgonBoard board) {
+    public long evaluate(AgonBoard board) {
         return 0;
     }
 }
