+++ agon/src/main/java/fr/univ/bordeaux/application/ai/heuristics/MobilityHeuristic.java
@@ -1,8 +1,10 @@
 package fr.univ.bordeaux.application.ai.heuristics;
 
+import fr.univ.bordeaux.agonCore.bitboard.AgonBoard;
+
 public class MobilityHeuristic implements Heuristics {
     @Override
-    public double evaluate(AgonBoard board) {
+    public long evaluate(AgonBoard board) {
         return 0;
     }
 }
